[WIP] refactor task creation in backend
